2010-02-05 01:22 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/filesys.c     * UINT -> unsigned int     * Rename: bRead -> fRead, bResult -> fResult, usError -> nResult, uiResult -> nResult     * Formatting.
  * include/hbapifs.h
  * src/rtl/filesys.c
  * src/rtl/philes.c
  * src/rtl/philesx.c
  * contrib/hbct/disk.c
  * contrib/xhb/xhbfunc.c
    * Drive letter type change from USHORT/BYTE to int.
    + Extra checks added to enforce positive value.
    % Optimized out fResult from hb_fsCurDirBuff(), plus related cleanups.
    ; Peer-review it please.

  * src/rtl/fstemp.c
    * Pacified warnings.

  * contrib/hbwin/Makefile
  + contrib/hbwin/wapi_wingdi_fontmem.c
    + Added WAPI_ADDFONTRESOURCEEX().
    + Added WAPI_REMOVEFONTRESOURCEEX().

+++ harbour/src/rtl/philesx.c
@@ -79,7 +79,7 @@ HB_FUNC( CURDRIVE )
 
       if( iDrive >= 0 )
       {
-         while( hb_fsChDrv( ( BYTE ) iDrive ) != 0 )
+         while( hb_fsChDrv( iDrive ) != 0 )
          {
             USHORT uiAction = hb_errRT_BASE_Ext1( EG_OPEN, 6001, "Operating system error",
                                                   HB_ERR_FUNCNAME, 0, EF_CANDEFAULT | EF_CANRETRY,
